리팩토링: 중복 코드 제거 위해 component 분리
