+++ src/components/carousel/reviewCarousel/Profile.js
@@ -1,0 +1,46 @@
+import styled from 'styled-components';
+
+const Profile = ({ userInfo }) => {
+  const { thumbNail, name, timeCreated } = userInfo;
+  return (
+    <>
+      <ProfileWrap>
+        <img src={thumbNail} alt="" />
+        <div className="info">
+          <strong>
+            <a href="">{name}</a>
+          </strong>
+          <span>{timeCreated}</span>
+        </div>
+      </ProfileWrap>
+    </>
+  );
+};
+
+export default Profile;
+
+const ProfileWrap = styled.div`
+  overflow: auto;
+  margin: 0.5rem 0;
+  img {
+    float: left;
+    margin-right: 0.5rem;
+    border-radius: 50%;
+  }
+  .info {
+    float: left;
+    width: calc(100% - 48px);
+    margin-top: 0.5rem;
+    a {
+      float: left;
+      height: 100%;
+      margin-top: -7px;
+      padding: 7px;
+      color: #2e7d32;
+    }
+    span {
+      float: right;
+      font-size: 12px;
+    }
+  }
+`;
