+++ src/components/main/MainContainer.js
@@ -7,10 +7,10 @@ import Footer from '../footer/Footer';
 const MainContainer = () => {
   return (
     <MainWrap>
-      <Carousel />
-      <Carousel />
+      <Carousel carouselInfo={DUMMY_LIST} />
+      <Carousel carouselInfo={DUMMY_LIST} />
       <SquareCarousel />
-      <ReviewCarousel />
+      <ReviewCarousel carouselInfo={DUMMY_REVIEWS} />
       <Footer />
     </MainWrap>
   );
@@ -25,3 +25,121 @@ const MainWrap = styled.div`
   margin: 0 auto;
   padding: 3rem 1.5rem 0;
 `;
+
+const DUMMY_LIST = [
+  {
+    user: {
+      thumbNail: 'http://placehold.it/40x40.png?text=A',
+      name: 'Harry',
+      timeCreated: '10초 전',
+    },
+    review: {
+      src: 'http://placehold.it/270x175.png?text=A',
+      type: 'VG',
+      storeName: '발우공양',
+      region: '서울',
+      starRating: '⭐⭐⭐⭐⭐',
+      contents:
+        'Buddhist temple cuisine in a clean and modern space, at 서울시종로구견지동715F. from the Choe Gae Sa Temple',
+    },
+  },
+  {
+    user: {
+      thumbNail: 'http://placehold.it/40x40.png?text=A',
+      name: 'Harry',
+      timeCreated: '10초 전',
+    },
+    review: {
+      src: 'http://placehold.it/270x175.png?text=A',
+      type: 'VG',
+      storeName: '발우공양',
+      region: '서울',
+      starRating: '⭐⭐⭐⭐⭐',
+      contents:
+        'Buddhist temple cuisine in a clean and modern space, at 서울시종로구견지동715F. from the Choe Gae Sa Temple',
+    },
+  },
+  {
+    user: {
+      thumbNail: 'http://placehold.it/40x40.png?text=A',
+      name: 'Harry',
+      timeCreated: '10초 전',
+    },
+    review: {
+      src: 'http://placehold.it/270x175.png?text=A',
+      type: 'VG',
+      storeName: '발우공양',
+      region: '서울',
+      starRating: '⭐⭐⭐⭐⭐',
+      contents:
+        'Buddhist temple cuisine in a clean and modern space, at 서울시종로구견지동715F. from the Choe Gae Sa Temple',
+    },
+  },
+  {
+    user: {
+      thumbNail: 'http://placehold.it/40x40.png?text=A',
+      name: 'Harry',
+      timeCreated: '10초 전',
+    },
+    review: {
+      src: 'http://placehold.it/270x175.png?text=A',
+      type: 'VG',
+      storeName: '발우공양',
+      region: '서울',
+      starRating: '⭐⭐⭐⭐⭐',
+      contents:
+        'Buddhist temple cuisine in a clean and modern space, at 서울시종로구견지동715F. from the Choe Gae Sa Temple',
+    },
+  },
+];
+
+const DUMMY_REVIEWS = [
+  {
+    user: {
+      thumbNail: 'http://placehold.it/40x40.png?text=A',
+      name: 'Harry',
+      timeCreated: '10초 전',
+    },
+    review: {
+      src: 'http://placehold.it/365x235.png?text=A',
+      type: 'VG',
+      storeName: '발우공양',
+      region: '서울',
+      starRating: '⭐⭐⭐⭐⭐',
+      contents:
+        'Buddhist temple cuisine in a clean and modern space, at 서울시종로구견지동715F. from the Choe Gae Sa Temple',
+    },
+  },
+  {
+    user: {
+      thumbNail: 'http://placehold.it/40x40.png?text=A',
+      name: 'Harry',
+      timeCreated: '10초 전',
+    },
+    review: {
+      src: 'http://placehold.it/365x235.png?text=A',
+      type: 'VG',
+      storeName: '발우공양',
+      region: '서울',
+      starRating: '⭐⭐⭐⭐⭐',
+      contents:
+        'Buddhist temple cuisine in a clean and modern space, at 서울시종로구견지동715F. from the Choe Gae Sa Temple',
+    },
+  },
+  {
+    user: {
+      thumbNail: 'http://placehold.it/40x40.png?text=A',
+      name: 'Harry',
+      timeCreated: '10초 전',
+    },
+    review: {
+      src: 'http://placehold.it/365x235.png?text=A',
+      type: 'VG',
+      storeName: '발우공양',
+      region: '서울',
+      starRating: '⭐⭐⭐⭐⭐',
+      contents:
+        'Buddhist temple cuisine in a clean and modern space, at 서울시종로구견지동715F. from the Choe Gae Sa Temple',
+    },
+  },
+];
